Raise a ValidationError corresponding to the first unsatisfied constraint during entity validation
This prevents the wrong error message to be displayed when one has a
vocabulary constraint for which a size constraint is also automatically added.
Without the break, an error relative to the latter constraint will be raised
whereas the user has actually specified the former. This is also consistent
with other checks done in the method (stop on first "failure").

Closes #202779.