[constraint] make all constraints accept a 'msg' argument specifying custom failure message
This implementation serializes the message along with other constraint
attributes, and as such doesn't need any particular migration since previously
serialized values are properly decoded and message is added as needed.

Along the way, FormatConstraint has been refactored and JSON serialization of
BoundaryConstraint and StaticVocabularyConstraint has been updated to keyword
args, as for other, so msg is transparently deserialized.

Closes #288874